Backend/python-flask/requirements.txt
Flask==2.2
Flask-SQLAlchemy==3.0.2
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.2
psycopg2-binary==2.9.1
python-dotenv==0.14.0
Werkzeug==2.2.0
itsdangerous==2.0.1